pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "myloware"
version = "0.2.0"
description = "Production-grade multi-agent AI orchestration platform for automated video production workflows"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}

authors = [
    {name = "Mylo James", email = "mylo.james114@gmail.com"}
]

keywords = [
    "ai",
    "multi-agent",
    "orchestration",
    "langgraph",
    "langchain",
    "video-production",
    "video-generation",
    "automation",
    "fastapi",
    "postgresql",
    "pgvector",
    "asmr",
    "content-creation",
    "ai-agents",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "pydantic-settings>=2.4.0,<3.0",
    "langgraph>=0.1.19",
    "langgraph-checkpoint-postgres>=2.0.25",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langsmith>=0.1.74",
    "redis>=5.0.0",
    "psycopg[binary]>=3.1.12",
    "psycopg-pool>=3.2.7",
    "sqlalchemy>=2.0.32",
    "alembic>=1.13.2",
    "structlog>=24.2.0",
    "python-dotenv>=1.0.1",
    "tenacity>=8.3.0",
    "httpx>=0.27.2",
    "prometheus-client>=0.21.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/mylo-james/myloware"
Documentation = "https://github.com/mylo-james/myloware/tree/main/docs"
Repository = "https://github.com/mylo-james/myloware"
Issues = "https://github.com/mylo-james/myloware/issues"
Changelog = "https://github.com/mylo-james/myloware/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.2",
    "mypy>=1.11.1",
    "ruff>=0.6.7"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "live_smoke: opt-in tests that hit live providers and require secrets",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.build.targets.wheel]
packages = [
    "apps/api",
    "apps/orchestrator",
    "apps/mcp_adapter",
    "adapters",
    "content",
    "cli",
    "infrastructure",
]

[project.scripts]
mw-py = "cli.main:main"
mw = "cli.main:main"
